chore: TntPlaceListener
