+++ src/main/java/fr/openmc/core/utils/MaterialUtils.java
@@ -84,4 +84,31 @@ public class MaterialUtils {
     public static boolean isCustomCrop(String namespace) {
         return CUSTOM_CROPS.contains(namespace);
     }
+
+    private static final Set<Material> CONTAINERS = Set.of(
+            Material.CHEST,
+            Material.CHEST_MINECART,
+            Material.SHULKER_BOX,
+            Material.WHITE_SHULKER_BOX,
+            Material.BLUE_SHULKER_BOX,
+            Material.BROWN_SHULKER_BOX,
+            Material.CYAN_SHULKER_BOX,
+            Material.GRAY_SHULKER_BOX,
+            Material.GREEN_SHULKER_BOX,
+            Material.LIME_SHULKER_BOX,
+            Material.MAGENTA_SHULKER_BOX,
+            Material.ORANGE_SHULKER_BOX,
+            Material.YELLOW_SHULKER_BOX,
+            Material.LIGHT_BLUE_SHULKER_BOX,
+            Material.LIGHT_GRAY_SHULKER_BOX,
+            Material.PINK_SHULKER_BOX,
+            Material.RED_SHULKER_BOX,
+            Material.PURPLE_SHULKER_BOX,
+            Material.BARREL,
+            Material.DECORATED_POT
+    );
+
+    public static boolean isContainers(Material type) {
+        return CONTAINERS.contains(type);
+    }
 }
